Update on the Review Comments
