+++ go.mod
@@ -1,4 +1,4 @@
-module beckn_onix
+module github.com/beckn/beckn-onix
 
 go 1.23.0
 
@@ -6,13 +6,14 @@ toolchain go1.23.7
 
 require (
 	github.com/stretchr/testify v1.10.0
-	golang.org/x/crypto v0.35.0
-	gopkg.in/yaml.v3 v3.0.1
+	golang.org/x/crypto v0.36.0
 )
+
+require gopkg.in/yaml.v3 v3.0.1 // indirect
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/sys v0.30.0 // indirect
+	golang.org/x/sys v0.31.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0
 )
